backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Praveen Kumar",email = "praveenkumar.e2022@vitstudent.ac.in"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "numpy (>=2.2.2,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "xgboost-cpu (>=2.1.4,<3.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
